+++ gdbgui/backend.py
@@ -32,8 +32,8 @@ from pygdbmi.gdbcontroller import GdbController
 BASE_PATH = os.path.dirname(os.path.realpath(__file__))
 PARENTDIR = os.path.dirname(BASE_PATH)
 sys.path.append(PARENTDIR)
-from gdbgui import htmllistformatter
-from gdbgui import __version__
+from gdbgui import htmllistformatter  # noqa
+from gdbgui import __version__  # noqa
 
 TEMPLATE_DIR = os.path.join(BASE_PATH, 'templates')
 STATIC_DIR = os.path.join(BASE_PATH, 'static')
@@ -50,6 +50,13 @@ if match is not None and int(match.groups()[0]) >= 16:
     # if mac OS version is 16 (sierra) or higher, need to set shell off due to
     # os's security requirements
     STARTUP_WITH_SHELL_OFF = True
+
+
+# create dictionary of signal names
+SIGNAL_NAME_TO_NUM = {}
+for n in dir(signal):
+    if n.startswith('SIG') and '_' not in n:
+        SIGNAL_NAME_TO_NUM[n.upper()] = getattr(signal, n)
 
 # Create flask application and add some configuration keys to be used in various callbacks
 app = Flask(__name__, template_folder=TEMPLATE_DIR, static_folder=STATIC_DIR)
@@ -191,6 +198,7 @@ def run_gdb_command(message):
             _gdb_state['gdb_controllers'].get(request.sid).write(cmd, read_response=False)
 
         except Exception as e:
+            print(e)
             err = traceback.format_exc()
             dbprint(traceback.format_exc())
             emit('error_running_gdb_command', {'message': err})
@@ -277,6 +285,7 @@ def gdbgui():
             'initial_binary_and_args': app.config['initial_binary_and_args'],
             'show_gdbgui_upgrades': app.config['show_gdbgui_upgrades'],
             'themes': THEMES,
+            'signals': SIGNAL_NAME_TO_NUM
         }
 
     return render_template('gdbgui.pug',
@@ -285,6 +294,18 @@ def gdbgui():
         interpreter=interpreter,
         initial_data=json.dumps(initial_data),
         themes=THEMES)
+
+
+@app.route('/send_signal_to_pid')
+def send_signal_to_pid():
+    signal_name = request.args.get('signal_name', '')
+    signal_num = SIGNAL_NAME_TO_NUM.get(signal_name.upper())
+    if signal is None:
+        raise ValueError('no such signal %s' % signal_name)
+
+    pid = int(request.args.get('pid'))
+    os.kill(pid, signal_num)
+    return jsonify({'message': 'sent signal %s (%s) to process id %s' % (signal_name, signal_num, str(pid))})
 
 
 @app.route('/shutdown')